- Centered display. - Adjusted colors.
git-svn-id: svn://127.0.0.1/wator/trunk@7 80941b80-f086-49b8-bfa8-edbbee4b5223

+++ wator.cpp
@@ -45,7 +45,7 @@ int WINAPI WinMain(HINSTANCE hinst,
 			}
 		}
 
-		iReturn = msg.wParam;
+		iReturn = static_cast<int>(msg.wParam);
 	}
 	catch(...)
 	{
